fix(frontend): share Chart.js registration to cut Sonar dups
Register dashboard charts once and use useCurrentOrgBillingCycleChart
so file preambles no longer share long duplicated new-code blocks.
