+++ src/components/dashboard/BundleUploadsChart.vue
@@ -2,27 +2,16 @@
 import type { ChartData, ChartOptions, Plugin } from 'chart.js'
 import type { TooltipClickHandler } from '~/services/chartTooltip'
 import { useDark } from '@vueuse/core'
-import {
-  BarController,
-  BarElement,
-  CategoryScale,
-  Chart,
-  LinearScale,
-  LineController,
-  LineElement,
-  PointElement,
-  Tooltip,
-} from 'chart.js'
 import { computed } from 'vue'
 import { Bar, Line } from 'vue-chartjs'
 import { useI18n } from 'vue-i18n'
 import { useRouter } from 'vue-router'
-import { useOrgBillingCycleChart } from '~/composables/useOrgBillingCycleChart'
+import { useCurrentOrgBillingCycleChart } from '~/composables/useOrgBillingCycleChart'
 import { createLegendConfig, createStackedChartScales } from '~/services/chartConfig'
 import { createTodayLineOptions, generateAppChartColors, getSafeChartHue } from '~/services/chartTodayLine'
 import { createTooltipConfig, todayLinePlugin, verticalLinePlugin } from '~/services/chartTooltip'
+import { registerDashboardCharts } from '~/services/dashboardChartRegister'
 import { generateMonthDays, getDaysInCurrentMonth } from '~/services/date'
-import { useOrganizationStore } from '~/stores/organization'
 import { createChartLegendItems } from './chartLegend'
 import ChartLegend from './ChartLegend.vue'
 
@@ -37,46 +26,25 @@ const props = defineProps({
   accumulated: { type: Boolean, default: false },
 })
 
+registerDashboardCharts()
+
 const isDark = useDark()
 const { t } = useI18n()
 const router = useRouter()
-const organizationStore = useOrganizationStore()
-const chartCycle = useOrgBillingCycleChart(
-  () => props.useBillingPeriod,
-  () => organizationStore.currentOrganization?.subscription_start,
-  () => organizationStore.currentOrganization?.subscription_end,
-)
+const chartCycle = useCurrentOrgBillingCycleChart(() => props.useBillingPeriod)
 const cycleStart = chartCycle.resolveCycleStart()
 const cycleEnd = chartCycle.resolveCycleEnd()
 const { todayLimit, transformDailySeries } = chartCycle
 
-// Create a reverse mapping from app name to app ID for tooltip clicks
-const appIdByLabel = computed(() => {
-  const mapping: Record<string, string> = {}
-  Object.entries(props.appNames as Record<string, string>).forEach(([appId, appName]) => {
-    mapping[appName] = appId
-  })
-  return mapping
+const tooltipClickHandler = computed<TooltipClickHandler>(() => {
+  const appIdByLabel: Record<string, string> = {}
+  for (const [appId, appName] of Object.entries(props.appNames as Record<string, string>))
+    appIdByLabel[appName] = appId
+  return {
+    onAppClick: (appId: string) => router.push(`/app/${appId}`),
+    appIdByLabel,
+  }
 })
-
-// Click handler for tooltip items - navigates to app detail page
-const tooltipClickHandler = computed<TooltipClickHandler>(() => ({
-  onAppClick: (appId: string) => {
-    router.push(`/app/${appId}`)
-  },
-  appIdByLabel: appIdByLabel.value,
-}))
-
-Chart.register(
-  Tooltip,
-  BarController,
-  BarElement,
-  LineController,
-  LineElement,
-  PointElement,
-  CategoryScale,
-  LinearScale,
-)
 
 function monthdays() {
   return generateMonthDays(props.useBillingPeriod, cycleStart, cycleEnd)
